Refactor DebugTheme: pull out reusable "FloatingDebugButton" component
